Python v3/requirements.txt
discord.py
requests
tweepy
python-dotenv
